chronicler/client/pyproject.toml
[project]
name = "chronicler-client"
version = "0.1.0"
description = "Chronicler Python Client"
authors = [{ name = "Chronicler Foundation", email = "chronicleraudit@gmail.com" }]
requires-python = ">=3.10"

dependencies = [
    "dspy",
    "web3",
    "python-dotenv",
    "pydantic>=2.0.0",
    "fastapi",
    "uvicorn",
    "httpx",
    "pytest",
    "ruff",
    "fastapi-mcp>=0.3.7",
]

[tool.uv]
package = true

[tool.setuptools]
packages = ["chronicler_client"]